debug similarity type, test task3

+++ project2/src/cbir.cpp
@@ -244,8 +244,8 @@ void readDB_rec(char *dir, char ***fileArr, int *max, int *numFile) {
 }
 
 int imgComparator(const void* p1, const void* p2) {
-  int img1Similarity = (*(Img **)p1)->getSimilarity();
-  int img2Similarity = (*(Img **)p2)->getSimilarity();
+  float img1Similarity = (*(Img **)p1)->getSimilarity();
+  float img2Similarity = (*(Img **)p2)->getSimilarity();
   // printf("comparing %d and %d\n", img1Similarity, img2Similarity);
   return img2Similarity-img1Similarity;
 }
